Read Among Us memory to enable automatic muting (#1)
The bot now reads Among Us' memory to determine which players are alive and dead and what the current state of the game is. It uses this information to determine which players should be muted, moved or able to speak at any given time.

The bot must now be updated every time the game is updated.

The bot must be run by the host of the lobby or it will fail to detect the end of meetings.

Despite these limitations, this is a much smoother experience and is much preferred by the people I've play tested it with. Thanks again to @lewis-weinberger, @HJeffery, @EdanMiles and @CrumpetsOnToast for their help.